+++ backend/src/main/java/com/arthur/lojaonline/repository/EnderecosRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface EnderecosRepository extends JpaRepository<Enderecos, Integer>{
+public interface EnderecosRepository extends JpaRepository<Enderecos, Long>{
     
     List<Enderecos> findByClienteId(Long clienteId);
